Revert "trying to rewrite day and event render logic"
This reverts commit faa23e241504ede04cc7ce383e937d481ce79096.
